backend/studio/backend/requirements.txt
fastapi
uvicorn
python-multipart
watchdog
openai
python-dotenv
requests
sentence-transformers
transformers
chromadb
nltk
spacy
Pillow
pytesseract
pdfplumber
pymupdf
opencv-python
numpy
scikit-image
matplotlib
pandas
scikit-learn
python-dateutil
regex
typing-extensions
